MM-24379: Adds e2e tests.
